agents/example_agent/requirements.txt
google-adk>=1.2.0
google-cloud-aiplatform[agent_engines,adk]>=1.88.0
httpx>=0.27.0
mcp>=1.0.0
